+++ src/main/java/com/example/surveyapp/domain/ai/moderation/application/facade/AiModerationFacade.java
@@ -1,7 +1,7 @@
 package com.example.surveyapp.domain.ai.moderation.application.facade;
 
-import com.example.surveyapp.domain.ai.moderation.domain.vo.AiModerationResult;
-import com.example.surveyapp.domain.ai.moderation.domain.vo.AiModerationSurveyResult;
+import com.example.surveyapp.domain.ai.moderation.domain.model.vo.AiModerationResult;
+import com.example.surveyapp.domain.ai.moderation.domain.model.vo.AiModerationSurveyResult;
 
 public interface AiModerationFacade {
     AiModerationResult checkNicknameModeration(String nickname);
