+++ src/main/java/com/example/surveyapp/domain/ai/moderation/application/facade/AiModerationFacadeImpl.java
@@ -1,9 +1,9 @@
 package com.example.surveyapp.domain.ai.moderation.application.facade;
 
 import com.example.surveyapp.domain.ai.moderation.application.AiModerationService;
-import com.example.surveyapp.domain.ai.moderation.domain.model.AiModerationTargetType;
-import com.example.surveyapp.domain.ai.moderation.domain.vo.AiModerationResult;
-import com.example.surveyapp.domain.ai.moderation.domain.vo.AiModerationSurveyResult;
+import com.example.surveyapp.domain.ai.moderation.domain.model.enums.AiModerationTargetType;
+import com.example.surveyapp.domain.ai.moderation.domain.model.vo.AiModerationResult;
+import com.example.surveyapp.domain.ai.moderation.domain.model.vo.AiModerationSurveyResult;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
